Flash to kotlin fixes for lists

+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/authorized/ConsoleCommandsListMessage.kt
@@ -13,17 +13,17 @@ open class ConsoleCommandsListMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		aliases = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readUTF()
 			aliases.add(item)
 		}
 		args = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readUTF()
 			args.add(item)
 		}
 		descriptions = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readUTF()
 			descriptions.add(item)
 		}
