+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/game/context/roleplay/havenbag/HavenBagFurnituresRequestMessage.kt
@@ -13,17 +13,17 @@ open class HavenBagFurnituresRequestMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		cellIds = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readVarShort().toInt()
 			cellIds.add(item)
 		}
 		funitureIds = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readInt().toInt()
 			funitureIds.add(item)
 		}
 		orientations = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readUnsignedByte().toInt()
 			orientations.add(item)
 		}
