+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/game/guild/GuildInformationsMembersMessage.kt
@@ -12,7 +12,7 @@ open class GuildInformationsMembersMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		members = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = GuildMember()
 			item.deserialize(stream)
 			members.add(item)
