+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/game/idol/IdolListMessage.kt
@@ -14,17 +14,17 @@ open class IdolListMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		chosenIdols = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readVarShort().toInt()
 			chosenIdols.add(item)
 		}
 		partyChosenIdols = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readVarShort().toInt()
 			partyChosenIdols.add(item)
 		}
 		partyIdols = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = ProtocolTypeManager.getInstance<PartyIdol>(stream.readUnsignedShort())
 			item.deserialize(stream)
 			partyIdols.add(item)
