+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/game/inventory/exchanges/ExchangeStartOkMountWithOutPaddockMessage.kt
@@ -12,7 +12,7 @@ open class ExchangeStartOkMountWithOutPaddockMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		stabledMountsDescription = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = MountClientData()
 			item.deserialize(stream)
 			stabledMountsDescription.add(item)
