+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/messages/game/progression/suggestion/ActivitySuggestionsMessage.kt
@@ -12,12 +12,12 @@ open class ActivitySuggestionsMessage : NetworkMessage() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		lockedActivitiesIds = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readVarShort().toInt()
 			lockedActivitiesIds.add(item)
 		}
 		unlockedActivitiesIds = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = stream.readVarShort().toInt()
 			unlockedActivitiesIds.add(item)
 		}
