+++ src/main/kotlin/fr/lewon/dofus/bot/sniffer/model/types/game/context/roleplay/breach/ExtendedBreachBranch.kt
@@ -14,7 +14,7 @@ open class ExtendedBreachBranch : BreachBranch() {
 	override fun deserialize(stream: ByteArrayReader) {
 		super.deserialize(stream)
 		rewards = ArrayList()
-		for (i in 0 until stream.readUnsignedShort()) {
+		for (i in 0 until stream.readUnsignedShort().toInt()) {
 			val item = BreachReward()
 			item.deserialize(stream)
 			rewards.add(item)
